icu_locale_core::extensions

Module unicode

Unicode Extensions provide information about user preferences in a given locale.

The main struct for this extension is Unicode which contains Keywords and Attributes.

Examples

use icu::locale::extensions::unicode::{attribute, key, value, Unicode};
use icu::locale::Locale;

let loc: Locale = "en-US-u-foobar-hc-h12".parse().expect("Parsing failed.");

assert_eq!(
    loc.extensions.unicode.keywords.get(&key!("hc")),
    Some(&value!("h12"))
);
assert!(loc
    .extensions
    .unicode
    .attributes
    .contains(&attribute!("foobar")));

Macros

attribute

A macro allowing for compile-time construction of valid Attribute subtags.

key

A macro allowing for compile-time construction of valid Key subtags.

subdivision_suffix

A macro allowing for compile-time construction of valid SubdivisionSuffix subtags.

value

A macro allowing for compile-time construction of valid Unicode Value subtag.

Structs

Attribute

An attribute used in a set of Attributes.

Attributes

A set of Attribute elements as defined in Unicode Extension Attributes.

Key

A key used in a list of Keywords.

Keywords

A list of Key-Value pairs representing functional information about locale’s internationalization preferences.

SubdivisionId

A Subivision Id as defined in Unicode Locale Identifier.

SubdivisionSuffix

A subdivision suffix used in SubdivisionId.

Unicode

Unicode Extensions provide information about user preferences in a given locale.

Value

A value used in a list of Keywords.