icu_locale_core::extensions::unicode

Struct Value

pub struct Value(/* private fields */);

A value used in a list of Keywords.

The value has to be a sequence of one or more alphanumerical strings separated by -. Each part of the sequence has to be no shorter than three characters and no longer than 8.

Examples

use icu::locale::extensions::unicode::{value, Value};
use writeable::assert_writeable_eq;

assert_writeable_eq!(value!("gregory"), "gregory");
assert_writeable_eq!(
    "islamic-civil".parse::<Value>().unwrap(),
    "islamic-civil"
);

// The value "true" has the special, empty string representation
assert_eq!(value!("true").to_string(), "");

Implementations

impl Value

pub const fn as_single_subtag(&self) -> Option<&Subtag>

Returns a reference to a single Subtag if the Value contains exactly one subtag, or None otherwise.

Examples

use core::str::FromStr;
use icu::locale::extensions::unicode::Value;

let value1 = Value::from_str("foo").expect("failed to parse a Value");
let value2 = Value::from_str("foo-bar").expect("failed to parse a Value");

assert!(value1.as_single_subtag().is_some());
assert!(value2.as_single_subtag().is_none());

pub fn into_single_subtag(self) -> Option<Subtag>

Destructs into a single Subtag if the Value contains exactly one subtag, or returns None otherwise.

Examples

use core::str::FromStr;
use icu::locale::extensions::unicode::Value;

let value1 = Value::from_str("foo").expect("failed to parse a Value");
let value2 = Value::from_str("foo-bar").expect("failed to parse a Value");

assert!(value1.into_single_subtag().is_some());
assert!(value2.into_single_subtag().is_none());

pub fn subtag_count(&self) -> usize

Returns the number of subtags in the Value.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};

let mut v = Value::default();
assert_eq!(v.subtag_count(), 0);
v.push_subtag(subtag!("foo"));
assert_eq!(v.subtag_count(), 1);

pub const fn new_empty() -> Self

Creates an empty Value, which corresponds to a “true” value.

Examples

use icu::locale::extensions::unicode::{value, Value};

assert_eq!(value!("true"), Value::new_empty());

pub fn is_empty(&self) -> bool

Returns true if the Value has no subtags.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};

let mut v = Value::default();
assert_eq!(v.is_empty(), true);

pub fn remove_subtag(&mut self, idx: usize) -> Option<Subtag>

Removes and returns the subtag at position index within the value, shifting all subtags after it to the left.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};
let mut v = Value::default();
v.push_subtag(subtag!("foo"));
v.push_subtag(subtag!("bar"));
v.push_subtag(subtag!("baz"));

assert_eq!(v.remove_subtag(1), Some(subtag!("bar")));
assert_eq!(v, "foo-baz");

pub fn get_subtag(&self, idx: usize) -> Option<&Subtag>

Returns a reference to a subtag at index.

Examples

use icu::locale::{extensions::unicode::Value, subtags::subtag};
let mut v = Value::default();
v.push_subtag(subtag!("foo"));
v.push_subtag(subtag!("bar"));
v.push_subtag(subtag!("baz"));

assert_eq!(v.get_subtag(1), Some(&subtag!("bar")));
assert_eq!(v.get_subtag(3), None);

impl Value

pub fn to_string(&self) -> String

Converts the given value to a String.

Under the hood, this uses an efficient [Writeable] implementation. However, in order to avoid allocating a string, it is more efficient to use [Writeable] directly.