icu_locale_core::preferences::extensions

Module unicode

A set of unicode extensions which correspond to preferences.

The module contains a set structs corresponding to Locale unicode extensions for which ICU4X provides implementations of preferences.

The macros in this module provide wrappers for creating preferences based on enums and structs.

Modules

errors

Errors related to parsing of Preferences.

keywords

A list of Preferences derived from Locale unicode extension keywords.

Macros

struct_keyword

Macro used to generate a preference keyword as a struct.