icu_locale_core::preferences
Macro define_preferences

macro_rules! define_preferences {
    (
        $(#[$doc:meta])*
        $([$derive_attrs:ty])?
        $name:ident,
        {
            $(
                $(#[$key_doc:meta])*
                $key:ident: $pref:ty
            ),*
        }
     ) => { ... };
}

A macro to facilitate generation of preferences struct.

The generated preferences struct provides methods for merging and converting between Locale and the preference bag. See preferences for use cases.

In the example below, the input argument is the generated preferences struct which can be auto-converted from a Locale, or combined from a Locale and Preferences Bag.

Examples

use icu::locale::{
    preferences::{
        define_preferences,
        extensions::unicode::keywords::HourCycle
    },
    locale,
};

define_preferences!(
    [Copy]
    NoCalendarFormatterPreferences,
    {
        hour_cycle: HourCycle
    }
);

struct NoCalendarFormatter {}

impl NoCalendarFormatter {
    pub fn try_new(prefs: NoCalendarFormatterPreferences) -> Result<Self, ()> {
        // load data and set struct fields based on the prefs input
        Ok(Self {})
    }
}

let loc = locale!("en-US");

let tf = NoCalendarFormatter::try_new(loc.into());