icu_locale_core::extensions::unicode

Struct Unicode

pub struct Unicode {
    pub keywords: Keywords,
    pub attributes: Attributes,
}

Unicode Extensions provide information about user preferences in a given locale.

A list of Unicode BCP47 U Extensions as defined in Unicode Locale Identifier specification.

Unicode extensions provide subtags that specify language and/or locale-based behavior or refinements to language tags, according to work done by the Unicode Consortium. (See RFC 6067 for details).

Examples

use icu::locale::extensions::unicode::{key, value};
use icu::locale::Locale;

let loc: Locale =
    "de-u-hc-h12-ca-buddhist".parse().expect("Parsing failed.");

assert_eq!(
    loc.extensions.unicode.keywords.get(&key!("ca")),
    Some(&value!("buddhist"))
);

Fields
keywords: Keywords

The key-value pairs present in this locale extension, with each extension key subtag associated to its provided value subtag.

attributes: Attributes

A canonically ordered sequence of single standalone subtags for this locale extension.

Implementations

impl Unicode

pub const fn new() -> Self

Returns a new empty map of Unicode extensions. Same as default(), but is const.

Examples

use icu::locale::extensions::unicode::Unicode;

assert_eq!(Unicode::new(), Unicode::default());

pub fn is_empty(&self) -> bool

Returns true if there list of keywords and attributes is empty.

Examples

use icu::locale::Locale;

let loc: Locale = "en-US-u-foo".parse().expect("Parsing failed.");

assert!(!loc.extensions.unicode.is_empty());

pub fn clear(&mut self)

Clears all Unicode extension keywords and attributes, effectively removing the Unicode extension.

Example

use icu::locale::Locale;

let mut loc: Locale =
    "und-t-mul-u-hello-ca-buddhist-hc-h12".parse().unwrap();
loc.extensions.unicode.clear();
assert_eq!(loc, "und-t-mul".parse().unwrap());

pub fn total_cmp(&self, other: &Self) -> Ordering

Returns an ordering suitable for use in BTreeSet.

The ordering may or may not be equivalent to string ordering, and it may or may not be stable across ICU4X releases.