icu_locale_core::preferences::extensions::unicode

Module keywords

A list of Preferences derived from Locale unicode extension keywords.

Structs

CurrencyType

A Unicode Currency Identifier defines a type of currency.

NumberingSystem

A Unicode Number System Identifier defines a type of number system.

RegionOverride

A Region Override specifies an alternate region to use for obtaining certain region-specific default values.

RegionalSubdivision

A Unicode Subdivision Identifier defines a regional subdivision used for locales.

TimeZoneShortId

A Unicode Timezone Identifier defines a timezone.

Enums

CalendarAlgorithm

A Unicode Calendar Identifier defines a type of calendar.

CollationCaseFirst

Collation parameter key for ordering by case.

CollationNumericOrdering

Collation parameter key for numeric handling.

CollationType

A Unicode Collation Identifier defines a type of collation (sort order).

CommonVariantType

A Unicode Variant Identifier defines a special variant used for locales.

CurrencyFormatStyle

A Unicode Currency Format Identifier defines a style for currency formatting.

EmojiPresentationStyle

A Unicode Emoji Presentation Style Identifier

FirstDay

A Unicode First Day Identifier defines the preferred first day of the week for calendar display.

HijriCalendarAlgorithm

Hijri Calendar sub-type

HourCycle

A Unicode Hour Cycle Identifier defines the preferred time cycle. Specifying “hc” in a locale identifier overrides the default value specified by supplemental time data for the region.

LineBreakStyle

A Unicode Line Break Style Identifier defines a preferred line break style corresponding to the CSS level 3 line-break option.

LineBreakWordHandling

A Unicode Line Break Word Identifier defines preferred line break word handling behavior corresponding to the CSS level 3 word-break option.

MeasurementSystem

A Unicode Measurement System Identifier defines a preferred measurement system.

MeasurementUnitOverride

A Measurement Unit Preference Override defines an override for measurement unit preference.

SentenceBreakSupressions

A Unicode Sentence Break Suppressions Identifier defines a set of data to be used for suppressing certain sentence breaks that would otherwise be found by UAX #14 rules.