icu_locale_core::preferences::extensions::unicode
Module keywords

A list of Preferences derived from Locale unicode extension keywords.

Structs

CurrencyType

A Unicode Currency Identifier defines a type of currency.

NumberingSystem

A Unicode Number System Identifier defines a type of number system.

RegionOverride

A Region Override specifies an alternate region to use for obtaining certain region-specific default values.

RegionalSubdivision

A Unicode Subdivision Identifier defines a regional subdivision used for locales.

TimeZoneShortId

A Unicode Timezone Identifier defines a timezone.